+++ CHANGELOG.md
@@ -23,6 +23,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Return string-keyed arrays from `GraphQL::getStandardDirectives()`, `GraphQL::getStandardTypes()` and `GraphQL::getStandardValidationRules()`
 - Move complexity related code from `FieldDefinition` to `QueryComplexity`
 - Exclude unused standard types from the schema
+- Require lazy type loader to return `Type` directly without an intermediary callable
 
 ### Added